migrate to typeScript & move folder
