+++ server/src/config/jest.js
@@ -1,6 +1,3 @@
-/* eslint-disable import/no-extraneous-dependencies */
-import jest from 'jest';
-
 const argv = process.argv.slice(2);
 
 // Watch unless on CI or explicitly running all tests
@@ -8,4 +5,5 @@ if (!process.env.CI) {
   argv.push('--watch');
 }
 
+// eslint-disable-next-line @typescript-eslint/no-unsafe-call
 jest.run(argv);
